+++ ros2_ws/src/rvr_node/sphero_sdk_wrapper/sphero_rvr_interface/sphero_rvr_mock.py
@@ -9,18 +9,18 @@ class SpheroRvrMock(metaclass=SpheroRvrInterface):
         """Close the connection to the RVR and put it to sleep"""
         pass
 
-    def on_will_sleep_notify(self, handler, timeout: float) -> None:
+    def on_will_sleep_notify(self, handler, timeout: int) -> None:
         """Run an action 10s before the RVR sleeps."""
         pass
     
-    def start_roll(self, speed: float, heading: float) -> None:
+    def start_roll(self, speed: int, heading: int) -> None:
         """Start roll at a specified speed and heading."""
         pass
 
-    def stop_roll(self, heading: float) -> None:
+    def stop_roll(self, heading: int) -> None:
         """Stop roll."""
         pass
 
-    def set_heading(self, heading: float) -> None:
+    def set_heading(self, heading: int) -> None:
         """Set heading."""
         pass
